Class-ified version control systems
