+++ .ruby-gemset
@@ -1,0 +1,1 @@
+rspec_profiling
