+++ lib/rspec_profiling.rb
@@ -5,6 +5,9 @@ require "rspec_profiling/version"
 require "rspec_profiling/run"
 require "rspec_profiling/collectors/csv"
 require "rspec_profiling/collectors/database"
+require "rspec_profiling/vcs/git"
+require "rspec_profiling/vcs/svn"
+require "rspec_profiling/vcs/git_svn"
 
 module RspecProfiling
   class Railtie < Rails::Railtie
