+++ lib/rspec_profiling/config.rb
@@ -5,6 +5,7 @@ module RspecProfiling
 
   def self.config
     @config ||= OpenStruct.new({
+      vcs: RspecProfiling::VCS::Git,
       collector:  RspecProfiling::Collectors::Database,
       db_path:    'tmp/rspec_profiling',
       table_name: 'spec_profiling_results',
